generic setter and getter, gpio init & reset
Use generic getter and setter functions for the direction and value
files of a gpio. Specific setter and getter for direction and value
inherit these functions.
Initialize a GPIO pin with the expected type and a path
saved within a struct instance of the gpio struct.
Reset a GPIO after the function completes.